+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## v1.3.13
+## v1.4.0
 
 ### Features
 
@@ -8,6 +8,8 @@
 - Add `download` route.
 - Add support for private S3 bucket.
 - Add presigned URL support for MinIO and AWS.
+- Add team based authorization matrix
+- Add `whoami` route.
 
 ## v1.3.12